extend constants

+++ headers/uki_input.h
@@ -51,10 +51,10 @@ bool uki_input_double_force_minmax(const char *msg, const char *repeat_msg, doub
 bool uki_input_double_force_constraint(const char *msg, const char *repeat_msg, double *value, uki_constraint_double_t constraint);
 
 // Для ввода строки в статический буфер
-bool uki_input_str(const char *msg, const char *err_msg, char *value, uint8_t buf_size);
-bool uki_input_str_force(const char *msg, const char *repeat_msg, char *value, uint8_t buf_size);
-bool uki_input_str_constraint(const char *msg, const char *err_msg, char *value, uint8_t buf_size, uki_constraint_str_t constraint);
-bool uki_input_str_force_constraint(const char *msg, const char *repeat_msg, char *value, uint8_t buf_size, uki_constraint_str_t constraint);
+bool uki_input_str(const char *msg, const char *err_msg, char *value, uint32_t buf_size);
+bool uki_input_str_force(const char *msg, const char *repeat_msg, char *value, uint32_t buf_size);
+bool uki_input_str_constraint(const char *msg, const char *err_msg, char *value, uint32_t buf_size, uki_constraint_str_t constraint);
+bool uki_input_str_force_constraint(const char *msg, const char *repeat_msg, char *value, uint32_t buf_size, uki_constraint_str_t constraint);
 
 // вариант сканфа, но в одну строку
 int uki_input_scanf_line(const char *msg, const char *fmt, ...);
